#DXP-451 Add cards list
