+++ blocks/carousel/carousel.css
@@ -1,4 +1,1 @@
 /* carousel csses */
-.carousel .card-content .rating {
-  display: flex;
-}
