+++ data/cards-list.json
@@ -1,0 +1,94 @@
+[
+  [
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A1kNj5aPzdL.jpg",
+      "text": "Hoffman",
+      "rating": 1,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/hoffman-down-filled-performance-fabric-living-room-sectional-sofa-couch-127-w-ecru-b071fmsynh.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/91/914PIFu8oeL.jpg",
+      "text": "Tufted Chair",
+      "rating": 2,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/leila-velvet-tufted-chair-44-w-navy-blue-b075x4vwhg.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A11i7JV5oxL.jpg",
+      "text": "Tufted Sofa",
+      "rating": 3,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/leila-tufted-living-room-sofa-couch-88-w-smoky-teal-b075x4vwf9.html"
+    }
+  ],
+  [
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/81/81oXIL6P58L.jpg",
+      "text": "Leila",
+      "rating": 4,
+      "price": "1550€",
+      "link": "/content/puresight/pages/products/leila-velvet-ottoman-34-w-blue-b075x543tw.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A1VU-cBgyYL.jpg",
+      "text": "Tufted Chair",
+      "rating": 5,
+      "price": "650€",
+      "link": "/content/puresight/pages/products/leila-tufted-chair-44-w-sandstone-b075x4vwfh.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A1vpJx61s8L.jpg",
+      "text": "Rivet Sloane",
+      "rating": 0,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/rivet-sloane-mid-century-modern-loveseat-with-tufted-back-64-2-w-storm-b07p8srzs1.html"
+    }
+  ],
+  [
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A1R13eYKKKL.jpg",
+      "text": "Rivet",
+      "rating": 1,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/rivet-sloane-mid-century-modern-sofa-with-tufted-back-79-9-w-storm-b07p7nvgfl.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A1e+te01d6L.jpg",
+      "text": "Tufted Chair",
+      "rating": 2,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/leila-tufted-chair-44-w-smoky-teal-b075x4vwfb.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/A1/A1D0wxYGe7L.jpg",
+      "text": "Marin Studded",
+      "rating": 3,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/marin-studded-chair-33-w-linen-b075x4n551.html"
+    }
+  ],
+  [
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/B1/B1d9hkLaFRS.jpg",
+      "text": "Hoffman",
+      "rating": 4,
+      "price": "1550€",
+      "link": "/content/puresight/pages/products/hoffman-down-filled-performance-fabric-loveseat-sofa-79-w-grey-tweed-b072jg3wsn.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/B1/B1QJINggHqS.jpg",
+      "text": "Hoffman",
+      "rating": 5,
+      "price": "650€",
+      "link": "/content/puresight/pages/products/hoffman-down-filled-performance-fabric-oversized-arm-chair-59-w-ecru-b071zlxzq7.html"
+    },
+    {
+      "img": "https://storage.googleapis.com/puresight-pim-images/original/91/91v4dFyZ1TL.jpg",
+      "text": "Tufted Sofa",
+      "rating": 0,
+      "price": "450€",
+      "link": "/content/puresight/pages/products/leila-tufted-velvet-living-room-sofa-couch-88-w-blue-b075x4f5lm.html"
+    }
+  ]
+]
